% Copyright (c) 2024, Avraham Adler All rights reserved
% SPDX-License-Identifier: MPL-2.0+

\name{MBBEFDLite-internal}
\title{Internal MBBEFDLite functions}
\description{Internal MBBEFDLite functions}
\details{These functions are not meant to be directly called by the user.}

\alias{checkgcb} % Checks for g and b if c is present.
\alias{c2gb}     % Converts c parameter to g and b.
\alias{findb}    % Used to converge to a b given a g and a mu.
\alias{dilog}    % Dilogarithm function used in Bernegger algorithm.
\alias{mugb}     % Return mu giveb g and b. Used in Bernegger algorithm.
\alias{cvgb}     % Return cv giveb g and b. Used in Bernegger algorithm.

\keyword{internal}
