% Copyright (c) 2024, Avraham Adler All rights reserved
% SPDX-License-Identifier: MPL-2.0+

\name{MBBEFDLite-package}
\alias{MBBEFDLite-package}
\encoding{UTF-8}
\docType{package}
\title{
\packageTitle{MBBEFDLite}
}
\description{
\packageDescription{MBBEFDLite}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MBBEFDLite}
\packageIndices{MBBEFDLite}
}
\author{
\packageAuthor{MBBEFDLite}

Maintainer: \packageMaintainer{MBBEFDLite}
}

\keyword{package}
\keyword{distribution}
