% Copyright (c) 2024, Avraham Adler All rights reserved
% SPDX-License-Identifier: MPL-2.0+

\name{MBBEFDLite-internal}
\title{Internal MBBEFDLite functions}
\description{Internal MBBEFDLite functions}
\details{These functions are not meant to be directly called by the user.}

\alias{checkgcb} % Checks for g and b if c is present
\alias{c2gb}     % Converts c parameter to g and b

\keyword{internal}
